t: Implement initial data seeding for products
This commit introduces a `CommandLineRunner` to populate the H2 database
with sample product data on application startup. This ensures that the
API has data to serve for development and testing purposes right from
the start.

Key changes:
- Added a `@Bean` of type `CommandLineRunner` in `ClothingApplication`
  to execute code after the application context is loaded.
  - Populated the database with three sample products (jacket, t-shirt,
    chinos) using the `ProductRepository`.
    - Corrected the instantiation of `Product` objects to use setters
      instead of a non-existent constructor, resolving a startup
      failure.
      - Ensured type safety for monetary values by using
	`BigDecimal.valueOf()` for product prices, aligning with the
	`Product` entity's definition.

+++ src/main/java/com/example/store/service/OrderService.java
@@ -37,7 +37,7 @@ public class OrderService {
             OrderItem item = new OrderItem();
             item.setProduct(cartItem.getProduct());
             item.setQuantity(cartItem.getQuantity());
-            item.setPriceAtPurchase(cartItem.getProduct().getPrice());
+            item.setPriceAtPurchase(cartItem.getProduct().getPrice().doubleValue());
             return item;
         }).collect(Collectors.toList());
